+++ vck/src/jvmTest/kotlin/at/asitplus/wallet/lib/jws/JweServiceJvmTest.kt
@@ -10,34 +10,53 @@ import at.asitplus.signum.indispensable.josef.JweHeader
 import at.asitplus.signum.indispensable.nativeDigest
 import at.asitplus.signum.indispensable.toJcaPublicKey
 import at.asitplus.signum.HazardousMaterials
+import at.asitplus.signum.indispensable.SecretExposure
+import at.asitplus.signum.indispensable.josef.JweAlgorithm.A128GCMKW
+import at.asitplus.signum.indispensable.josef.JweAlgorithm.A128KW
+import at.asitplus.signum.indispensable.josef.JweAlgorithm.A192GCMKW
+import at.asitplus.signum.indispensable.josef.JweAlgorithm.A192KW
+import at.asitplus.signum.indispensable.josef.JweAlgorithm.A256GCMKW
+import at.asitplus.signum.indispensable.josef.JweAlgorithm.A256KW
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
+import at.asitplus.signum.indispensable.symmetric.AuthCapability
+import at.asitplus.signum.indispensable.symmetric.KeyType
+import at.asitplus.signum.indispensable.symmetric.NonceTrait
+import at.asitplus.signum.indispensable.symmetric.SymmetricEncryptionAlgorithm
+import at.asitplus.signum.indispensable.symmetric.SymmetricEncryptionAlgorithm.Companion.AES_128
+import at.asitplus.signum.indispensable.symmetric.SymmetricEncryptionAlgorithm.Companion.AES_192
+import at.asitplus.signum.indispensable.symmetric.SymmetricEncryptionAlgorithm.Companion.AES_256
+import at.asitplus.signum.indispensable.symmetric.SymmetricKey
+import at.asitplus.signum.indispensable.symmetric.hasDedicatedMacKey
+import at.asitplus.signum.indispensable.symmetric.isAuthenticated
+import at.asitplus.signum.indispensable.symmetric.randomKey
+import at.asitplus.signum.indispensable.symmetric.secretKey
 import at.asitplus.signum.supreme.hazmat.jcaPrivateKey
 import at.asitplus.signum.supreme.sign.EphemeralKey
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
 import com.benasher44.uuid.uuid4
 import com.nimbusds.jose.*
+import com.nimbusds.jose.crypto.AESDecrypter
+import com.nimbusds.jose.crypto.AESEncrypter
 import com.nimbusds.jose.crypto.ECDHDecrypter
 import com.nimbusds.jose.crypto.ECDHEncrypter
 import com.nimbusds.jose.jwk.JWK
 import io.kotest.core.spec.style.FreeSpec
+import io.kotest.engine.runBlocking
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
-import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.JsonPrimitive
 import java.security.interfaces.ECPrivateKey
 import java.security.interfaces.ECPublicKey
 
-@OptIn(HazardousMaterials::class)
+@OptIn(HazardousMaterials::class, SecretExposure::class)
 class JweServiceJvmTest : FreeSpec({
 
-    val configurations: List<Configuration> =
-        listOf(
-            Configuration(SECP_256_R_1, listOf(A128CBC_HS256, A128GCM)),
-            Configuration(SECP_384_R_1, listOf(A192CBC_HS384, A192GCM)),
-            Configuration(SECP_521_R_1, listOf(A256CBC_HS512, A256GCM)),
-        )
+    val ecdhesConfiguration = listOf(
+        EcdhesConfiguration(SECP_256_R_1, listOf(A128CBC_HS256, A128GCM)),
+        EcdhesConfiguration(SECP_384_R_1, listOf(A192CBC_HS384, A192GCM)),
+        EcdhesConfiguration(SECP_521_R_1, listOf(A256CBC_HS512, A256GCM)),
+    )
 
-    configurations.forEach { config ->
+    ecdhesConfiguration.forEach { config ->
         val ephemeralKey = EphemeralKey {
             ec {
                 curve = config.curve
@@ -55,7 +74,7 @@ class JweServiceJvmTest : FreeSpec({
         val randomPayload = uuid4().toString()
 
         config.encryption.forEach { encryptionMethod ->
-            "${config.curve}, ${encryptionMethod}:" - {
+            "${config.curve}, ${encryptionMethod}" - {
                 "Encrypted object from ext. library can be decrypted with int. library" {
                     val libJweHeader =
                         JWEHeader.Builder(JWEAlgorithm(jweAlgorithm.identifier), encryptionMethod.joseAlgorithm)
@@ -89,9 +108,70 @@ class JweServiceJvmTest : FreeSpec({
             }
         }
     }
+
+    val symmetricConfiguration = listOf(
+        SymmetricConfiguration(A128KW, listOf(A128CBC_HS256, A128GCM)),
+        SymmetricConfiguration(A192KW, listOf(A192CBC_HS384, A192GCM)),
+        SymmetricConfiguration(A256KW, listOf(A256CBC_HS512, A256GCM)),
+        SymmetricConfiguration(A128GCMKW, listOf(A128CBC_HS256, A128GCM)),
+        SymmetricConfiguration(A192GCMKW, listOf(A192CBC_HS384, A192GCM)),
+        SymmetricConfiguration(A256GCMKW, listOf(A256CBC_HS512, A256GCM)),
+    )
+
+    symmetricConfiguration.forEach { config ->
+        runBlocking {
+            val ephemeralKey = (config.algorithm as JweAlgorithm.Symmetric).algorithm.randomKey()
+            require(ephemeralKey is SymmetricKey.Integrated)
+
+            val jvmEncrypter = AESEncrypter(ephemeralKey.secretKey.getOrThrow())
+            val jvmDecrypter = AESDecrypter(ephemeralKey.secretKey.getOrThrow())
+
+            val encrypter = EncryptJweSymmetric(ephemeralKey)
+            val decrypter = DecryptJweSymmetric(ephemeralKey)
+            val randomPayload = uuid4().toString()
+
+            config.encryption.forEach { encryptionMethod ->
+                "${config.algorithm.identifier}, ${encryptionMethod}" - {
+                    "Encrypted object from ext. library can be decrypted with int. library" {
+                        val libJweHeader =
+                            JWEHeader.Builder(JWEAlgorithm(config.algorithm.identifier), encryptionMethod.joseAlgorithm)
+                                .type(JOSEObjectType("something"))
+                                .build()
+                        val libJweObject = JWEObject(libJweHeader, Payload(randomPayload))
+                            .apply { encrypt(jvmEncrypter) }
+                        val encryptedJwe = libJweObject.serialize()
+
+                        val parsedJwe = JweEncrypted.deserialize(encryptedJwe).getOrThrow()
+                        val result = decrypter(parsedJwe).getOrThrow()
+                        result.payload shouldBe randomPayload
+                    }
+
+                    "Encrypted object from int. library can be decrypted with ext. library" {
+                        val encrypted = encrypter(
+                            JweHeader(
+                                algorithm = config.algorithm,
+                                encryption = encryptionMethod,
+                            ),
+                            randomPayload,
+                        ).getOrThrow().serialize()
+
+                        val parsed = JWEObject.parse(encrypted).shouldNotBeNull()
+
+                        parsed.decrypt(jvmDecrypter)
+                        parsed.payload.toBytes().decodeToString() shouldBe randomPayload
+                    }
+                }
+            }
+        }
+    }
 })
 
-private data class Configuration(
+private data class SymmetricConfiguration(
+    val algorithm: JweAlgorithm,
+    val encryption: Collection<JweEncryption>,
+)
+
+private data class EcdhesConfiguration(
     val curve: ECCurve,
     val encryption: Collection<JweEncryption>,
 )